// packages/typescript/src/generated/ClientFrame.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Request } from "./Request.js";
import type { RpcError } from "./RpcError.js";
import type { JsonValue } from "./serde_json/JsonValue.js";

export type ClientFrame = { "type": "request", id: string, request: Request, } | { "type": "callback_result", id: string, result: JsonValue | null, error: RpcError | null, };